% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{all_member_sizes}
\alias{all_member_sizes}
\title{Extract a vector of all member sizes}
\usage{
all_member_sizes(steel_grade, member_type, list_reference_tables)
}
\arguments{
\item{steel_grade}{steel_grade [\eqn{N/{mm}^2}], categorical: 'S355' or 'S275'}

\item{member_type}{member_type, categorical: 'UC' or 'UB'}

\item{list_reference_tables}{List of reference tables}
}
\value{
Vector of all member sizes [ height (mm) x width (mm) x mass (kg/m) ]
}
\description{
Extract a vector of all member sizes for specified steel grade (S355/S275) and member type (UC/UB)
}
