% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{axial_compression_force}
\alias{axial_compression_force}
\title{Calculate the axial compression force}
\usage{
axial_compression_force(DL = 1, LL = 1, L, AF, theta = 90,
  spacing = 6, Lcry = 12.7, Lcrz = 12.7, steel_grade = "S355",
  member_type = "UB", alpha_T = 1.2e-05, delta_T = 10, k_T = 0.8,
  E = 210, IL = 50)
}
\arguments{
\item{DL}{Dead load / self-weight of member [\eqn{kN/m}]}

\item{LL}{Live load / imposed load [\eqn{kN/m}]}

\item{L}{Total length of member [\eqn{m}]}

\item{AF}{Axial compression force of member per meter [\eqn{kN/m}]}

\item{theta}{Angle to wall [\eqn{deg}]}

\item{spacing}{spacing [\eqn{m}]}

\item{Lcry}{critical length major axis [\eqn{m}]}

\item{Lcrz}{critical length minor axis [\eqn{m}]}

\item{steel_grade}{steel_grade [\eqn{N/{mm}^2}], categorical: 'S355' or 'S275'}

\item{member_type}{member_type, categorical: 'UC' or 'UB'}

\item{alpha_T}{Thermal coef. of expansion [\eqn{degC}]}

\item{delta_T}{Change in temperature from the Installation temperature [\eqn{degC}]}

\item{k_T}{Coefficient Of Temperature Effect [dimensionless]}

\item{E}{Young's Modulus of Elasticity [\eqn{GPa}]}

\item{IL}{Accidental Impact Load [\eqn{kN/m}]}
}
\value{
\eqn{N_{ed}} Axial compression force [\eqn{kN}]
}
\description{
Compute Axial Compression Force, \eqn{N_{ed}} [\eqn{kN}], for member without including Temperature effect.
Used as trial for the top level strut where temperature changes could not be neglected.
As well can be used to calculate final \eqn{N_{ed}} for struts from low levels of excavation, where temperature effect could be neglected.
}
\details{
First of all function check which combination govern in ULS (Ultimate Limit State) without including Temperature load, TL [\eqn{kN}].
Then include TL calculations for Load Combinations applying partial factors based on the Table A1.2(B), EN1990-2002, p53
Compare maximum from ULS and ALS to define which mistake could govern.
}
