% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{axial_compression_force_given_member}
\alias{axial_compression_force_given_member}
\title{Calculate the axial compression force for a given member size}
\usage{
axial_compression_force_given_member(isTopLevel = T, Ned_no_TL = 6987,
  member_size, alpha_T = 1.2e-05, delta_T = 10, k_T = 0.8, E = 210,
  list_reference_tables)
}
\arguments{
\item{isTopLevel}{Is member located at top level? [boolean]}

\item{alpha_T}{Thermal coef. of expansion [\eqn{degC}]}

\item{delta_T}{Change in temperature from the Installation temperature [\eqn{degC}]}

\item{k_T}{Coefficient Of Temperature Effect [dimensionless]}

\item{E}{Young's Modulus of Elasticity [\eqn{GPa}]}

\item{list_reference_tables}{List of reference tables}
}
\value{
\itemize{
  \item \eqn{N_{ed}} Axial compression force [\eqn{kN}]
  \item \eqn{TL} Temperature Load [\eqn{kN}]
}
}
\description{
Compute Axial Compression Force, \eqn{N_{ed}} [\eqn{kN}], for member without including Temperature effect.
Used as trial for the top level strut where temperature changes could not be neglected.
As well can be used to calculate final \eqn{N_{ed}} for struts from low levels of excavation, where temperature effect could be neglected.
}
\details{
First of all function check which combination govern in ULS (Ultimate Limit State) without including Temperature load, TL [\eqn{kN}].
Then include TL calculations for Load Combinations applying partial factors based on the Table A1.2(B), EN1990-2002, p53
Compare maximum from ULS and ALS to define which mistake could govern.
}
