% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{axial_compression_force_Ned}
\alias{axial_compression_force_Ned}
\title{critical_length_major_axis_Lcry (m)}
\usage{
axial_compression_force_Ned(DL = 1, LL = 1, L, AF, theta = 90,
  spacing = 6, Lcry = 12.7, Lcrz = 12.7, steel_grade = "S355",
  member_type = "UB", alpha_T = 1.2e-05, delta_T = 10, k_T = 0.8,
  E = 210, IL = 50)
}
\arguments{
\item{x}{A description of the parameter 'x'. The
description can span multiple lines.}

\item{y}{A description of the parameter 'y'.}
}
\value{
Lcrz critical_length_major_axis_Lcrz [m]
}
\description{
This is the description.
}
\details{
These are further details.
}
\section{A Custom Section}{


Text accompanying the custom section.
}

\examples{
add_numbers(1, 2) ## returns 3

}
