% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_TL}
\alias{calculate_TL}
\title{Calculate the temperature load}
\usage{
calculate_TL(alpha_T = 1.2e-05, delta_T = 10, k_T = 0.8, E = 210,
  A)
}
\arguments{
\item{alpha_T}{Thermal coefficient of expansion [degC]}

\item{delta_T}{Change in temperature from the Installation temperature [degC]}

\item{k_T}{Coefficient Of temperature effect [dimensionless]}

\item{E}{Young's Modulus of Elasticity [GPa]}

\item{A}{Sectional area from table for given member size [cm2]}
}
\value{
\eqn{TL} Temperature load [kN]
}
\description{
Calculate Temperature Load as a function of a surface changes of temperature.
Usually used for calculation of Axial Compression Force for the top level member. \deqn{TL = \alpha_T \, \delta_T \, k_T  \, E \, A}
}
