% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_overall_buckling_resistance_about_yy_axis}
\alias{check_overall_buckling_resistance_about_yy_axis}
\title{critical_length_major_axis_Lcry (m)}
\usage{
check_overall_buckling_resistance_about_yy_axis(trial_member_size,
  member_type, steel_grade, k, L, E)
}
\arguments{
\item{x}{A description of the parameter 'x'. The
description can span multiple lines.}

\item{y}{A description of the parameter 'y'.}
}
\value{
Lcrz critical_length_major_axis_Lcrz [m]
}
\description{
This is the description.
}
\details{
These are further details.
}
\section{A Custom Section}{


Text accompanying the custom section.
}

\examples{
add_numbers(1, 2) ## returns 3

}
