% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{combined_vertical_load}
\alias{combined_vertical_load}
\title{USL vs ALS verical load combinations}
\usage{
combined_vertical_load(DL, LL, AL)
}
\arguments{
\item{DL}{Dead load / self-weight of member [\eqn{kN/m}]}

\item{LL}{Live load / imposed load [\eqn{kN/m}]}

\item{AL}{Accidental Impact Load [\eqn{kN/m}]}
}
\value{
combined_vertical_load [\eqn{kN/m}]
}
\description{
USL vs ALS verical load combinations
}
\details{
Calculation steps are as follows:
\enumerate{
  \item \eqn{ULS: F=(1.35\,DL +1.5\,LL+1.5\,TL)}
  \item \eqn{ALS: F=(1.0\,DL +0.7\,LL+1.0\,AL)}
  \item \eqn{ALS: F=(1.0\,DL +0.6\,LL+1.0\,AL+0.5\,TL)}
}
}
