% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_output_table}
\alias{compute_output_table}
\title{Export output table to Excel file}
\usage{

  compute_output_table(file_name = "tables/input/output_processed_table.xlsx")
}
\arguments{
\item{file_name}{Path and file name of the output table}
}
\value{
None
}
\description{
Export output table to Excel file.
}
