% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{convert_member_dimensions_string_to_elements}
\alias{convert_member_dimensions_string_to_elements}
\title{Convert individual member dimensions to a string}
\usage{
convert_member_dimensions_string_to_elements(s)
}
\arguments{
\item{s}{String of the member dimensions}
}
\value{
\itemize{
  \item \eqn{h} Member height [\eqn{mm}]
  \item \eqn{b} Member width [\eqn{mm}]
  \item \eqn{m} Member weight [\eqn{kg/m}]
}
}
\description{
Convert individual member dimensions to a string.
}
