% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{convert_member_dimensions_to_string}
\alias{convert_member_dimensions_to_string}
\title{Convert the member size individual dimensions to a standard string}
\usage{
convert_member_dimensions_to_string(h, b, m)
}
\arguments{
\item{h}{Member height [\eqn{mm}]}

\item{b}{Member width [\eqn{mm}]}

\item{m}{Member mass [\eqn{kg/m}]}
}
\value{
String of the member dimensions
}
\description{
Generate a combined string from given three individual elements, separated by "x".
}
