% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{critical_length_major_axis_Lcry}
\alias{critical_length_major_axis_Lcry}
\title{Calculate the critical length along major \eqn{y} axis}
\usage{
critical_length_major_axis_Lcry(L, Lkp, Lsp)
}
\arguments{
\item{L}{Total length of member [m]}

\item{Lkp}{Length to king post [m]}

\item{Lsp}{Length from splays [m]}
}
\value{
\eqn{L_{cry}} critical_length_major_axis_Lcry [m]
}
\description{
This is the description.
}
