% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{critical_length_major_axis_y}
\alias{critical_length_major_axis_y}
\title{Calculate the critical length along major \eqn{y} axis}
\usage{
critical_length_major_axis_y(L, Lkp, Lsp)
}
\arguments{
\item{L}{Total length of member [\eqn{m}]}

\item{Lkp}{Length to king post [\eqn{m}]}

\item{Lsp}{Length from splays [\eqn{m}]}
}
\value{
\eqn{L_{cry}} Critical length along major \eqn{y} axis [\eqn{m}]
}
\description{
Calculate the critical length along major \eqn{y} axis, \eqn{L_{cry}} [\eqn{m}]
}
