% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{critical_length_minor_axis_Lcrz}
\alias{critical_length_minor_axis_Lcrz}
\title{Calculate the critical length along major \eqn{z} axis}
\usage{
critical_length_minor_axis_Lcrz(L, Lkp, Lsp)
}
\arguments{
\item{L}{Total length of member [m]}

\item{Lkp}{Length to king post [m]}

\item{Lsp}{Length from splays [m]}
}
\value{
\eqn{L_{crz}} critical_length_major_axis_Lcrz [m]
}
\description{
This is the description.
}
