% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{critical_length_minor_axis_z}
\alias{critical_length_minor_axis_z}
\title{Calculate the critical length along major \eqn{z} axis}
\usage{
critical_length_minor_axis_z(L, Lkp, Lsp)
}
\arguments{
\item{L}{Total length of member [\eqn{m}]}

\item{Lkp}{Length to king post [\eqn{m}]}

\item{Lsp}{Length from splays [\eqn{m}]}
}
\value{
\eqn{L_{crz}} Critical length along major \eqn{z} axis [\eqn{m}]
}
\description{
Calculate the critical length along major \eqn{z} axis, \eqn{L_{crz}} [\eqn{m}]
}
