% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{effective_length_of_member}
\alias{effective_length_of_member}
\title{Calculate the effective length of member}
\usage{
effective_length_of_member(k, L)
}
\arguments{
\item{k}{Effective lengh coefficient [dimensionless]}

\item{L}{Length of strut between restraints [\eqn{m}]}
}
\value{
\eqn{L_e} Effective length of strut [\eqn{m}]
}
\description{
Calculate the effective length of member, \eqn{L_e} [\eqn{m}].
}
