% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{effective_length_of_strut}
\alias{effective_length_of_strut}
\title{Calculate the effective length of strut}
\usage{
effective_length_of_strut(k, L)
}
\arguments{
\item{\eqn{k}}{Effective lengh coefficient [dimensionless]}

\item{\eqn{L}}{Length of strut between restraints [mm]}
}
\value{
\eqn{L_e} Effective length of strut [mm]
}
\description{
This is the description.
}
