% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extract_member_dimensions}
\alias{extract_member_dimensions}
\title{Extract dimensions from reference table}
\usage{
extract_member_dimensions(h, b, m, member_type, list_reference_tables)
}
\arguments{
\item{h}{Member height [\eqn{mm}]}

\item{b}{Member width [\eqn{mm}]}

\item{m}{Member mass [\eqn{kg/m}]}

\item{member_type}{Member type, 'UB' or 'UC'}

\item{list_reference_tables}{List of reference tables}
}
\value{
\itemize{
  \item \eqn{A} Area of section [\eqn{cm^2}]
  \item \eqn{tw} Thickness of web [\eqn{mm}]
  \item \eqn{tf} Thickness of flange [\eqn{mm}]
  \item \eqn{Iyy} Second moment of area axis \eqn{y-y} [\eqn{{cm}^4}]
  \item \eqn{sh} Depth of section [\eqn{mm}]
  \item \eqn{sb} Width of section [\eqn{mm}]
  \item \eqn{Izz} Second moment of area axis \eqn{z-z} [\eqn{{cm}^4}]
}
}
\description{
Function that looks into the Blue Book \url{https://www.steelforlifebluebook.co.uk/} for dimensions and properties.
}
