% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{first_order_bending_moment}
\alias{first_order_bending_moment}
\title{Calculate first order bending moment about major axis \eqn{y-y} [\eqn{kN.m}]}
\usage{
first_order_bending_moment(combined_vertical_load, L)
}
\arguments{
\item{combined_vertical_load}{combined_vertical_load [\eqn{kN/m}]}

\item{L}{Length of strut between restraints [\eqn{m}]}
}
\value{
\eqn{M^I_{Ed}} First order bending moment [\eqn{kN.m}]
}
\description{
Calculate first order bending moment about major axis \eqn{y-y} [\eqn{kN.m}]
}
\details{
Calculated as \deqn{M^I_{Ed} = 0.08\,F\,L_{cr,y}^2}
}
