% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fy}
\alias{fy}
\title{\eqn{f_y} Yield strength}
\usage{
fy(tw, tf, steel_grade)
}
\arguments{
\item{tw}{Thickness of the web [mm]}

\item{tf}{Thickness of the flange [mm]}
}
\value{
\eqn{f_y} Yield strength [N/mm2]
}
\description{
This is the description.
}
\examples{
fy(tw=47.6, tf=77, steel_grade="S355")

}
