% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{imperfection_factor_alpha_yy}
\alias{imperfection_factor_alpha_yy}
\title{Calculate the imperfection factor \eqn{\alpha_{yy}} for rolled section [dimensionless]}
\usage{
imperfection_factor_alpha_yy(h, b, tf)
}
\arguments{
\item{h}{Member height [mm]}

\item{b}{Member width [mm]}

\item{tf}{thickness of the flange [mm]}
}
\value{
\eqn{\alpha_{yy}} Imperfection factor for y-y axis
}
\description{
Calculate the imperfection factor \eqn{\alpha_{yy}} for rolled section [dimensionless]
}
