% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{imperfection_factor_alpha_zz}
\alias{imperfection_factor_alpha_zz}
\title{Calculate the imperfection factor \eqn{\alpha_{zz}} for rolled section [dimensionless]}
\usage{
imperfection_factor_alpha_zz(h, b, tf)
}
\arguments{
\item{h}{Member height [mm]}

\item{b}{Member width [mm]}

\item{tf}{thickness of the flange [mm]}
}
\value{
\eqn{\alpha_{zz}} Imperfection factor for z-z axis
}
\description{
This is the description.
}
