% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{imperfection_factor_yy}
\alias{imperfection_factor_yy}
\title{Calculate the imperfection factor \eqn{\alpha_{yy}} for rolled section [dimensionless]}
\usage{
imperfection_factor_yy(h, b, tf)
}
\arguments{
\item{h}{Member height [\eqn{mm}]}

\item{b}{Member width [\eqn{mm}]}

\item{tf}{thickness of the flange [\eqn{mm}]}
}
\value{
\eqn{\alpha_{yy}} Imperfection factor for \eqn{y-y} axis [dimensionless]
}
\description{
Calculate the imperfection factor \eqn{\alpha_{yy}} for rolled section [dimensionless].
}
