% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{imperfection_factor_zz}
\alias{imperfection_factor_zz}
\title{Calculate the imperfection factor \eqn{\alpha_{zz}} for rolled section}
\usage{
imperfection_factor_zz(h, b, tf)
}
\arguments{
\item{h}{Member height [\eqn{mm}]}

\item{b}{Member width [\eqn{mm}]}

\item{tf}{thickness of the flange [\eqn{mm}]}
}
\value{
\eqn{\alpha_{zz}} Imperfection factor for \eqn{z-z} axis [dimensionless]
}
\description{
Calculate the imperfection factor \eqn{\alpha_{zz}} for rolled section.
}
