% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{import_reference_BlueBook_tables}
\alias{import_reference_BlueBook_tables}
\title{Import Reference BlueBook Tables from Excel files}
\usage{
import_reference_BlueBook_tables()
}
\arguments{
\item{None}{}
}
\value{
List of 4 BlueBook reference tables
}
\description{
Import Reference BlueBook Tables from Excel files.
}
