% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{max_compressive_axial_force_in_chords}
\alias{max_compressive_axial_force_in_chords}
\title{Maximum compressive axial force in the chords}
\usage{
max_compressive_axial_force_in_chords(k, L, A, n, Ad, Lch, E, h0, Ned)
}
\arguments{
\item{k}{Coefficient of length as function of wall rigidity [dimensionless]}

\item{L}{Length between two restrains [\eqn{m}]}

\item{n}{Number of lacing planes, default [\eqn{n=2}]}

\item{Ad}{Section area of diagonal (lacing), [\eqn{cm^2}]}

\item{Lch}{Length of chord of betwen restrains (lace points) [\eqn{m}]}

\item{E}{Young modulus [\eqn{GPa} or \eqn{GN/m^2}]}

\item{h0}{Distance between centroids of chords [\eqn{m}]}

\item{Ned}{Axial compression Force [\eqn{kN}]}

\item{Ieff}{Effective second moment of area [\eqn{{mm}^4}]}

\item{Sv}{Shear stiffness for K-shape lacing [\eqn{kN}]}
}
\value{
\eqn{N_{ch,E_d}} Maximum compressive axial force in the chords [\eqn{kN}]
}
\description{
Determine maximum compressive axial force in the chords at mid-length of the strut, \eqn{N_{ch,E_d}} [\eqn{kN}]
}
\details{
Calculation steps are as follows:
\enumerate{
  \item Effective length \eqn{L_e = k \, L} [\eqn{mm}], where \eqn{L} is the length between two vertical supports of laced struts.
  \item Effective second moment of area \deqn{I_{eff}=0.5 \, h_0 \, A_{ch}}  [\eqn{{mm}^4}]. \eqn{I_{eff}}, where \eqn{h_0} [\eqn{cm}] is the distance between centroids of the chords and \eqn{A_{ch}}, [\eqn{{cm}^4}] is the cross-sectional area of one chord.
  \item Shear stiffness for K-shape lacing \eqn{S_v} [\eqn{kN}], where \eqn{d} is the lenth of diagonals \eqn{d = sqrt{ {h_0}^2 + L_{ch} }} [\eqn{mm}].
  \item Calculate the Euler buckling load \eqn{N_{cr,ch}} [\eqn{kN}] \deqn{ N_{cr,ch}=\frac{\pi^2 \, E \, I} {{L_e}^2} }, where \eqn{L_e} is the effective length between two vertical supports.
  \item Compute the second order bending moment \eqn{{M_{E_d}}^{II}} [\eqn{kN.m}].
  \item Output maximum compressive axial force in the chords \eqn{N_{ch,E_d}} [\eqn{kN}] \deqn{ N_{ch,E_d} = \frac{N_{E_d}}{2} + \frac{M_{E_d} \, h_0 \, A}{ 2 \, I_{eff}} }
}
}
