% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{member_size_to_string}
\alias{member_size_to_string}
\title{Convert the member size individual dimensions to a standard string}
\usage{
member_size_to_string(h, b, m)
}
\arguments{
\item{h}{Member height [mm]}

\item{b}{Member width [mm]}

\item{m}{Member mass [kg]}
}
\value{
String of the member dimensions
}
\description{
This is the description.
}
\examples{
member_size_to_string(h=610, b=229, m=140)

}
