% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plastic_resistance_of_cross_section_to_compression}
\alias{plastic_resistance_of_cross_section_to_compression}
\title{\eqn{N_{pl,R_d}} Plastic resistance of the cross-section to compression}
\usage{
plastic_resistance_of_cross_section_to_compression(A, fy)
}
\arguments{
\item{A}{Cross-section area of the strut [\eqn{cm^2}]}

\item{fy}{Yield strength [\eqn{kN/{mm}^2}]}
}
\value{
\eqn{N_{pl,R_d}} Plastic resistance of the cross-section to compression [kN]
}
\description{
This is the description.
}
