% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_input_table}
\alias{read_input_table}
\title{Read input table from given Excel file}
\usage{
read_input_table(file_name = "tables/input/trial1_kotik.xlsx")
}
\value{
Input table
}
\description{
Read input table from given Excel file.
}
