% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{second_order_moment}
\alias{second_order_moment}
\title{Calculate the second order moment}
\usage{
second_order_moment(L, Ned, Sv, Ncr)
}
\arguments{
\item{L}{Length of strut between restraints [mm]}

\item{Ned}{Axial_compression_force_Ned [kN]}

\item{Sv}{Shear stiffness for K-shape lacing}

\item{Ncr}{Euler buckling load from check #2 global zz [kN]}
}
\value{
\eqn{M_{E_d}} Second order moment [kN.m]
}
\description{
This is the description.
}
