% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{shear_force_at_support}
\alias{shear_force_at_support}
\title{Shear force at support calculation}
\usage{
shear_force_at_support(DL, LL, L, AL)
}
\arguments{
\item{DL}{Dead load / self-weight of member [\eqn{kN/m}]}

\item{LL}{Live load / imposed load [\eqn{kN/m}]}

\item{L}{Length of strut between restraints [\eqn{m}]}

\item{AL}{Accidental Impact Load [\eqn{kN/m}]}
}
\value{
\eqn{V_{Ed}} Shear force at support [\eqn{kN}]
}
\description{
Generate Shear force at support \eqn{V_{Ed}} [\eqn{kN}].
}
