% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{slenderness_reduction_factor}
\alias{slenderness_reduction_factor}
\title{Calculate the slenderness reduction factor}
\usage{
slenderness_reduction_factor(alpha, lambda_bar)
}
\arguments{
\item{alpha}{- check 1: imperfection_factor_alpha_yy for rolled section [dimentionless]}

\item{lambda_bar, }{Relative slenderness [dimentionless]}

\item{alpha}{- check 2 & 3: imperfection_factor_alpha_zz for rolled section [dimentionless]}
}
\value{
\eqn{X} Slenderness reduction factor [dimentionless]
}
\description{
This is the description.
}
