% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{temperature_load}
\alias{temperature_load}
\title{Calculate the temperature load}
\usage{
temperature_load(alpha_T = 1.2e-05, delta_T = 10, k_T = 0.8, E = 210, A = 94.4)
}
\arguments{
\item{alpha_T}{Thermal coefficient of expansion [\eqn{degC}]}

\item{delta_T}{Change in temperature from the Installation temperature [\eqn{degC}]}

\item{k_T}{Coefficient Of temperature effect [dimensionless]}

\item{E}{Young's Modulus of Elasticity [\eqn{GPa} or \eqn{GN/m2}]}

\item{A}{Sectional area from table for given member size [\eqn{cm^2}]}
}
\value{
TL Temperature load [\eqn{kN}]
}
\description{
Calculate Temperature Load as a function of a surface changes of temperature, TL [\eqn{kN}].
Usually used for calculation of Axial Compression Force for the top level member. \deqn{TL = \alpha_T \, \delta_T \, k_T  \, E \, A}
}
