% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{trial_member_size}
\alias{trial_member_size}
\title{Determine member size [ height (mm) x width (mm) x mass (kg/m) ]}
\usage{
trial_member_size(Lcry, Lcrz, Ned, steel_grade, member_type)
}
\arguments{
\item{Lcry}{critical length major axis [m]}

\item{Lcrz}{critical length minor axis [m]}

\item{Ned}{Axial compression force [kN]}

\item{steel}{grade, steel_grade [\eqn{N/mm^2}] - categorical: S355/S275}

\item{member}{type, member_type (categorical: UC/UB)}
}
\value{
Member size [ height (mm) x width (mm) x mass (kg/m) ]
}
\description{
This is the description.
}
\examples{
trial_member_size(Lcry=12.7, Lcrz=12.7, Ned=4926, steel_grade="S275", member_type="UC")

}
