% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{yield_strength}
\alias{yield_strength}
\title{Calculate the yield strength}
\usage{
yield_strength(tw, tf, steel_grade)
}
\arguments{
\item{tw}{Thickness of the web [\eqn{mm}]}

\item{tf}{Thickness of the flange [\eqn{mm}]}

\item{steel_grade}{steel_grade [\eqn{N/{mm}^2}], categorical: 'S355' or 'S275'}
}
\value{
\eqn{f_y} Yield strength [\eqn{N/{mm}^2}]
}
\description{
Calculate the yield strength, \eqn{f_y} [\eqn{N/{mm}^2}]
}
\examples{
yield_strength(tw=47.6, tf=77, steel_grade="S355")

}
