% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_TL}
\alias{calculate_TL}
\title{TL, temperature load [kN]}
\usage{
calculate_TL(alpha_T = 1.2e-05, delta_T = 10, k_T = 0.8, E = 210,
  A)
}
\arguments{
\item{alpha_T}{Thermal coefficient of expansion [degC]}

\item{delta_T}{Change in temperature from the Installation temperature [degC]}

\item{k_T}{Coefficient Of temperature effect [dimensionless]}

\item{E}{Young's Modulus of Elasticity [GPa]}

\item{A}{Sectional area from table for given member size [cm2]}
}
\value{
TL Temperature load [kN]
}
\description{
This is the description.
}
\examples{
calculate_TL <- function(alpha_T=0.000012, delta_T=10, k_T=0.8, E=210, A=200)

}
