% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_all_member_sizes}
\alias{check_all_member_sizes}
\title{Generate a table with all member sizes and apply all checks on each of them}
\usage{
check_all_member_sizes(steel_grade = "S355", member_type = "UB",
  k = 0.8, L = 12.5, E = 210, h0 = 1000, Lch = 1000, Ad = 1140,
  n = 2, isTopLevel = T, alpha_T = 1.2e-05, delta_T = 10,
  k_T = 0.8, Ned_no_TL = 6987, LL, AL, Lcry, ml,
  strut_name = "strut #",
  base_file_name = "tables/input/all_member_sizes_checked",
  export_xlsx = T)
}
\arguments{
\item{steel_grade}{steel_grade [\eqn{N/{mm}^2}], categorical: 'S355' or 'S275'}

\item{member_type}{member_type, categorical: 'UC' or 'UB'}

\item{k}{Coefficient [dimensionless]}

\item{L}{Total length of member [\eqn{m}]}

\item{E}{Young's Modulus of Elasticity [\eqn{GPa}]}

\item{h0}{Distance between centroids of chords [\eqn{mm}]}

\item{Lch}{Length of chord [\eqn{mm}]}

\item{Ad}{Section area of diagonal (lacing), [\eqn{cm^2}]}

\item{n}{Number of planes of lacing, default [\eqn{n=2}]}

\item{isTopLevel}{Is member located at top level? [boolean]}

\item{alpha_T}{Thermal coef. of expansion [\eqn{degC}]}

\item{delta_T}{Change in temperature from the Installation temperature [\eqn{degC}]}

\item{k_T}{Coefficient Of Temperature Effect [dimensionless]}

\item{Ned_no_TL}{Axial Compressional Force without Temperature Load [\eqn{kN}]}

\item{LL}{Live load / imposed load [\eqn{kN/m}]}

\item{AL}{Accidental Impact Load [\eqn{kN/m}]}

\item{Lcry}{critical length about major axis [\eqn{m}]}

\item{ml}{Lacing weight [\eqn{kN/m}]}

\item{strut_name}{Strut name}

\item{export_xlsx}{Boolean to export Excel spreadsheet or not [T/F]}

\item{file_name}{Path and file name of the output table}
}
\value{
\itemize{
  \item \eqn{df} Dataframe containing relevant input and all computed data
  \item \eqn{optimal_member_size} Optimal member size [ height (mm) x width (mm) x mass (kg/m) ]
  \item \eqn{optimal_TL} Optimal Temperature Load [\eqn{kN}]
  \item \eqn{optimal_Ned} Optimal @param Ned_no_TL Axial Compressional Force with Temperature Load [\eqn{kN}]
}
}
\description{
Generate a table with all member sizes and apply all checks on each of them
}
