% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_local_buckling_resistance_about_zz_axis}
\alias{check_local_buckling_resistance_about_zz_axis}
\title{Perform check #3, calculating the local buckling resistance of struts about z-z axis}
\usage{
check_local_buckling_resistance_about_zz_axis(trial_member_size,
  member_type, steel_grade, k, L, E)
}
\arguments{
\item{trial_member_size}{Trial member size}

\item{member_type}{member_type (categorical: UC/UB)}

\item{steel_grade}{steel_grade [\eqn{N/mm^2}] - categorical: S355/S275}

\item{k}{Coefficient [dimensionless]}

\item{L}{Total length of member [m]}

\item{E}{Young's Modulus of Elasticity [GPa]}
}
\value{
\eqn{N_{b,Rd,X}} Local buckling resistance of struts about z-z axis [kN]
}
\description{
This is the description.
}
