% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_overall_buckling_resistance_about_yy_axis}
\alias{check_overall_buckling_resistance_about_yy_axis}
\title{Perform check #1, calculating the overall buckling resistance of member about major \eqn{y-y} axis}
\usage{
check_overall_buckling_resistance_about_yy_axis(trial_member_size,
  member_type, steel_grade, k, L, E, list_reference_tables)
}
\arguments{
\item{trial_member_size}{Trial member size}

\item{member_type}{member_type, categorical: 'UC' or 'UB'}

\item{steel_grade}{steel_grade [\eqn{N/{mm}^2}], categorical: 'S355' or 'S275'}

\item{k}{Coefficient [dimensionless]}

\item{L}{Total length of member [\eqn{m}]}

\item{E}{Young's Modulus of Elasticity [\eqn{GPa}]}

\item{list_reference_tables}{List of reference tables}
}
\value{
\itemize{
  \item \eqn{N_{b,Rd,X}} Overall buckling resistance of struts about major y-y axis [\eqn{kN}]
  \item \eqn{N_{b,R_d,X}}
  \item \eqn{f_y}
  \item \eqn{N_{pl,R_d}}
  \item \eqn{N_{cr,X}}
  \item \eqn{N_{cr,X}}
  \item \eqn{{\bar{\lambda_X}}}
  \item \eqn{\alpha_{yy}}
  \item \eqn{X}
}
}
\description{
Calculate the overall buckling resistance of member about \eqn{y-y} axis, based on EC3 Approach. \deqn{L_e=kL} [\eqn{mm}]
where \eqn{L} is the critical length for buckling about major axis \eqn{y-y}
Steps of the check performed for laced struts:
\enumerate{
  \item Plastic resistance of the cross-section to compression [\eqn{kN}] \deqn{N_{pl,R_d}= 2(A \, fy)}
  \item The Euler buckling load [\eqn{kN}] \deqn{N_{cr,X}=\frac{\pi^2 \, E \, I_{yy}}{{L_e}^2}}
  \item Relative slenderness [dimensionless] \deqn{ \bar{\lambda_X} = \sqrt{ \frac{N_{pl,R_d}}{N_{cr,X}} } }
  \item Calculate \eqn{\Phi_X} parameter for slenderness reduction factor \deqn{ \Phi_X = 0.5 \left[ 1 + \alpha \left( \bar{\lambda_X}-0.2 \right) + {\bar{\lambda_X}}^2 \right] }
  \item Slenderness reduction factor [dimensionless] \deqn{ X_x = \frac{1}{ \Phi_X+\sqrt{{\Phi_X}^2-{\bar{\lambda_X}}^2} } }
  \item Output overall buckling resistance of the struts about \eqn{y-y} axis [\eqn{kN}] \deqn{ N_{b,R_d,X}=X_X \, N_{pl,R_d} }
}
The partial factors \eqn{\gamma_M} that are applied to resistance of members to instability:	\eqn{\gamma_{M_1} = 1}
}
