% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_output_table}
\alias{compute_output_table}
\title{Export output table to Excel file}
\usage{
compute_output_table(
  file_name = "tables/input/output_processed_table.xlsx",
  export_xlsx = T
)
}
\arguments{
\item{file_name}{Path and file name of the output table}

\item{export_xlsx}{Boolean to export Excel spreadsheet or not [T/F]}
}
\value{
None
}
\description{
Export output table to Excel file.
}
