% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{effective_second_moment_of_area}
\alias{effective_second_moment_of_area}
\title{Calculate the effective second moment of area}
\usage{
effective_second_moment_of_area(h0, A)
}
\arguments{
\item{h0}{Distance between centroids of chords [m]}

\item{A}{Cross-section area of strut [cm2]}
}
\description{
This is the description.
}
