% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Euler_buckling_load}
\alias{Euler_buckling_load}
\title{Calculate the Euler buckling load}
\usage{
Euler_buckling_load(Le, E, I)
}
\arguments{
\item{Le}{Effective length of strut [\eqn{mm}]}

\item{E}{Young modulus [\eqn{MPa} or \eqn{MN/m^2}]}

\item{I}{- check 1: \eqn{I_{yy}}, second moment of area Axis \eqn{y-y} [\eqn{{cm}^4}]. Check 2: \eqn{I_{eff}}, Effective second moment of area [\eqn{{mm}^4}]. Check 3: \eqn{I_{eff}} or \eqn{I_{zz}} [\eqn{{mm}^4}]}
}
\value{
\eqn{N_{cr}} Euler buckling load [\eqn{kN}]
}
\description{
Calculate the Euler buckling load [\eqn{kN}] \deqn{ N_{cr,ch}=\frac{\pi^2 \, E \, I}{{L_e}^2} }
}
