% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fy}
\alias{fy}
\title{Calculate the yield strength}
\usage{
fy(tw, tf, steel_grade)
}
\arguments{
\item{tw}{Thickness of the web [mm]}

\item{tf}{Thickness of the flange [mm]}

\item{steel_grade}{steel_grade [\eqn{N/mm^2}] - categorical: S355/S275}
}
\value{
\eqn{f_y} Yield strength [N/mm2]
}
\description{
Calculate the yield strength
}
\examples{
fy(tw=47.6, tf=77, steel_grade="S355")

}
