% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{max_compressive_axial_force_in_chords}
\alias{max_compressive_axial_force_in_chords}
\title{Maximum compressive axial force in the chords}
\usage{
max_compressive_axial_force_in_chords(
  trial_member_size,
  member_type,
  steel_grade,
  k,
  L,
  n,
  Ad,
  Lch,
  E,
  h0,
  Ned,
  list_reference_tables,
  isTopLevel,
  DL,
  LL,
  AL,
  TL,
  Lcry
)
}
\arguments{
\item{trial_member_size}{Trial member size}

\item{member_type}{member_type, categorical: 'UC' or 'UB'}

\item{steel_grade}{steel_grade [\eqn{N/{mm}^2}], categorical: 'S355' or 'S275'}

\item{k}{Coefficient of length as function of wall rigidity [dimensionless]}

\item{L}{Length between two restraints [\eqn{m}]}

\item{n}{Number of lacing planes, default [\eqn{n=2}]}

\item{Ad}{Section area of diagonal (lacing), [\eqn{cm^2}]}

\item{Lch}{Length of chord of betwen restrains (lace points) [\eqn{m}]}

\item{E}{Young modulus [\eqn{GPa} or \eqn{GN/m^2}]}

\item{h0}{Distance between centroids of chords [\eqn{m}]}

\item{Ned}{Axial compression Force [\eqn{kN}]}

\item{list_reference_tables}{List of reference tables}

\item{isTopLevel}{Is member located at top level? [boolean]}

\item{DL}{Dead load / self-weight of member [\eqn{kN/m}]}

\item{LL}{Live load / imposed load [\eqn{kN/m}]}

\item{AL}{Accidental Impact Load [\eqn{kN/m}]}

\item{TL}{Temperature load [\eqn{kN/m}]}

\item{Lcry}{critical length about major axis [\eqn{m}]}
}
\value{
\itemize{
  \item \eqn{N_{ch,E_d}} Maximum compressive axial force in the chords [\eqn{kN}]
  \item \eqn{S_v}
  \item \eqn{N_{cr,ch}}
  \item \eqn{{M_{E_d}}}
}
}
\description{
Determine maximum compressive axial force in the chords at mid-length of the strut, \eqn{N_{ch,E_d}} [\eqn{kN}]
}
\details{

}
