% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{maximum_shear_force_in_the_lacing}
\alias{maximum_shear_force_in_the_lacing}
\title{Calculate the maximum shear force in the lacing (for a laced strut subject to a compressive axial force only)}
\usage{
maximum_shear_force_in_the_lacing(MEd, L)
}
\arguments{
\item{MEd}{Second order moment [kN.m]}

\item{L}{Length of strut between restraints [mm]}
}
\value{
\eqn{V_{E_d}} Maximum shear force in the lacing (for a laced strut subject to a compressive axial force only)
}
\description{
This is the description.
}
