% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{member_dimensions}
\alias{member_dimensions}
\title{Extract dimensions from reference table}
\usage{
member_dimensions(h, b, m, member_type)
}
\arguments{
\item{h}{Member height [mm]}

\item{b}{Member width [mm]}

\item{m}{Member mass [kg]}

\item{member_type}{Member type, UB or UC}
}
\value{
\itemize{
  \item \eqn{A} Area of section [cm2]
  \item \eqn{tw} Thickness of web [mm]
  \item \eqn{tf} Thickness of flange [mm]
  \item \eqn{Iyy} Second moment of area axis y-y [cm4]
  \item \eqn{sh} Depth of section [mm]
  \item \eqn{sb} Width of section [mm]
}
}
\description{
This is the description.
}
\examples{
l1 <- member_dimensions(h=610, b=229, m=140, member_type="UB")
l2 <- member_dimensions(h=356, b=406, m=1299, member_type="UC")
l1$A; l2$tw

}
