% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{member_size_string_to_elements}
\alias{member_size_string_to_elements}
\title{Convert individual member dimensions to a string}
\usage{
member_size_string_to_elements(s)
}
\arguments{
\item{s}{String of the member dimensions}
}
\value{
\itemize{
  \item \eqn{h} Member height [mm]
  \item \eqn{b} Member width [mm]
  \item \eqn{m} Member weight [kg]
}
}
\description{
This is the description.
}
\examples{
member_size_string_to_elements("610 x 229 x 140")

}
