% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{overall_buckling_resistance_about_axis}
\alias{overall_buckling_resistance_about_axis}
\title{Calculate the overall buckling resistance of the struts about the axis}
\usage{
overall_buckling_resistance_about_axis(X, N_pl_Rd)
}
\arguments{
\item{X}{Slenderness reduction factor [dimentionless]}

\item{N_pl_Rd}{Plastic resistance of the cross-section to compression [kN]}
}
\value{
\eqn{N_{b,R_d}} Overall buckling resistance of the struts about the axis [kN]
}
\description{
This is the description.
}
