% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{relative_slenderness}
\alias{relative_slenderness}
\title{Calculate the relative slenderness}
\usage{
relative_slenderness(N_pl_Rd, Ncr)
}
\arguments{
\item{N_pl_Rd}{Plastic resistance of the cross-section to compression [kN]}

\item{Ncr}{Euler buckling load [kN]}
}
\value{
\eqn{\bar{\lambda}} Relative slenderness [dimentionless]
}
\description{
This is the description.
}
