% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{second_order_bending_moment}
\alias{second_order_bending_moment}
\title{Calculate the second order bending moment}
\usage{
second_order_bending_moment(L, Ned, Sv, Ncr, isTopLevel, DL, LL, IL, TL,
  Lcry)
}
\arguments{
\item{L}{Length of strut between restraints [\eqn{mm}]}

\item{Ned}{axial_compression_force [\eqn{kN}]}

\item{Sv}{Shear stiffness for K-shape lacing [\eqn{kN}]}

\item{Ncr}{Euler buckling load from check #2 global zz [\eqn{kN}]}

\item{isTopLevel}{Is member located at top level? [boolean]}

\item{DL}{Dead load / self-weight of member [\eqn{kN/m}]}

\item{LL}{Live load / imposed load [\eqn{kN/m}]}

\item{IL}{Accidental Impact Load [\eqn{kN/m}]}

\item{TL}{Temperature load [\eqn{kN/m}]}

\item{Lcry}{critical length about major axis [\eqn{m}]}
}
\value{
\eqn{M_{E_d}} Second order moment [\eqn{kN.m}]
}
\description{
Compute the second order bending moment, \eqn{M_{E_d}} [\eqn{kN.m}].
The maximum bending moment, including the bow imperfection and the second order effects, calculated as:
\deqn{ M_{E_d} = \frac{ N_{E_d} \, e_0 \, + \, {M_{E_d}}^I }{ 1 - \frac{N_{E_d}}{N_{cr,Y}} - \frac{N_{E_d}}{S_v} } }
}
