% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{shear_force_at_support}
\alias{shear_force_at_support}
\title{Shear force at support calculation}
\usage{
shear_force_at_support(isTopLevel, DL, LL, TL, Lcry)
}
\arguments{
\item{isTopLevel}{Is member located at top level? [boolean]}

\item{DL}{Dead load / self-weight of member [\eqn{kN/m}]}

\item{LL}{Live load / imposed load [\eqn{kN/m}]}

\item{TL}{Temperature load [\eqn{kN/m}]}

\item{Lcry}{critical length about major axis [\eqn{m}]}
}
\value{
\eqn{V_{Ed}} Shear force at support [\eqn{kN}]
}
\description{
Generate Shear force at support \eqn{V_{Ed}} [\eqn{kN}].
}
