% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{trial_member_size}
\alias{trial_member_size}
\title{Determine member size}
\usage{
trial_member_size(
  Lcry,
  Lcrz,
  Ned,
  steel_grade,
  member_type,
  list_reference_tables
)
}
\arguments{
\item{Lcry}{critical length major axis [\eqn{m}]}

\item{Lcrz}{critical length minor axis [\eqn{m}]}

\item{Ned}{Axial compression force [\eqn{kN}]}

\item{steel_grade}{steel_grade [\eqn{N/{mm}^2}], categorical: 'S355' or 'S275'}

\item{member_type}{member_type, categorical: 'UC' or 'UB'}

\item{list_reference_tables}{List of reference tables}
}
\value{
Member size [ height (mm) x width (mm) x mass (kg/m) ]
}
\description{
Find optimized designation [ height (mm) x width (mm) x mass (kg/m) ] (also called member size) for given Axial Compression Force and critical length for major and minor axis.
Searching into the tables based on the 'Compression' tables of the Blue Book \url{https://www.steelforlifebluebook.co.uk/}
}
